# backendtestpro/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BackendTestPro API",
    "version": "1.0.0",
    "description": "Backend və CTF/Lab funksionallıqları üçün tam API sənədi."
  },
  "paths": {
...
    "/run-test": {"post": {"summary": "Test başlat", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TestRequest"}}}}, "parameters": [{"name": "lang", "in": "query", "schema": {"type": "string"}}], "responses": {"200": {"description": "Test nəticəsi"}}}},
    "/report": {"post": {"summary": "Rapor", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TestRequest"}}}}, "parameters": [{"name": "lang", "in": "query", "schema": {"type": "string"}}], "responses": {"200": {"description": "Rapor"}}}},
...
  },
  "components": {
    "schemas": {
...
      "TestRequest": {"type": "object", "properties": {"target": {"type": "string"}}}
    }
  }
}